Extract shared and small Cloud API models
