+++ crates/clickhouse-cloud-api/src/models.rs
@@ -1,8 +1,8 @@
 //! Typed models for ClickHouse Cloud API schemas.
 //!
 //! Derived from the OpenAPI specification and kept in step with it by the drift
-//! analyzer, which parses this file — every model struct, enum and type alias
-//! must be declared here as literal source.
+//! analyzer, which parses the private module tree rooted at this facade. Every
+//! model struct, enum and type alias remains literal source in that tree.
 //!
 //! Request models are strict and response models have every field `Option<T>`; a
 //! schema used in both directions appears twice, as `{Name}` and
@@ -125,6 +125,71 @@ macro_rules! discriminated_union {
         }
     };
 }
+
+mod activity;
+mod api_keys;
+mod byoc;
+mod invitations;
+mod members;
+mod organization_private_endpoints;
+mod organizations;
+mod quotas;
+mod rbac;
+mod scim;
+mod shared;
+mod udfs;
+
+pub use activity::{Activity, ActivityActortype, ActivityKeyupdatetype, ActivityType};
+pub use api_keys::{
+    ApiKey, ApiKeyHashData, ApiKeyPatchRequest, ApiKeyPatchRequestState, ApiKeyPostRequest,
+    ApiKeyPostRequestState, ApiKeyPostResponse, ApiKeyState,
+};
+pub use byoc::{
+    ByocConfig, ByocConfigCloudprovider, ByocConfigRegionid, ByocConfigState,
+    ByocInfrastructurePatchRequest, ByocInfrastructurePostRequest,
+    ByocInfrastructurePostRequestRegionid,
+};
+pub use invitations::{
+    Invitation, InvitationPostRequest, InvitationPostRequestRole, InvitationRole,
+};
+pub use members::{Member, MemberPatchRequest, MemberPatchRequestRole, MemberRole};
+pub use organization_private_endpoints::{
+    OrganizationCloudRegionPrivateEndpointConfig, OrganizationPatchPrivateEndpoint,
+    OrganizationPatchPrivateEndpointCloudprovider, OrganizationPatchPrivateEndpointRegion,
+    OrganizationPrivateEndpoint, OrganizationPrivateEndpointCloudprovider,
+    OrganizationPrivateEndpointRegion, OrganizationPrivateEndpointsPatch,
+};
+pub use organizations::{Organization, OrganizationPatchRequest};
+pub use quotas::{OrganizationQuota, OrganizationQuotaQuotacode, OrganizationQuotaScope};
+pub use rbac::{
+    RBACPolicy, RBACPolicyAllowdeny, RBACPolicyCreateRequest, RBACPolicyCreateRequestAllowdeny,
+    RBACPolicyTags, RBACPolicyTagsResponse, RBACPolicyTagsRolev2, RBACRole, RBACRoleType,
+    RoleCreateRequest, RoleUpdateRequest,
+};
+pub use scim::{
+    ScimAuthenticationScheme, ScimBooleanFeature, ScimEnterpriseManager, ScimEnterpriseUser,
+    ScimGroup, ScimGroupListResponse, ScimGroupMember, ScimGroupMeta, ScimGroupPostRequest,
+    ScimGroupPutRequest, ScimListResponse, ScimPatchOp, ScimPatchOperation, ScimPatchOperationOp,
+    ScimResourceType, ScimResourceTypeListResponse, ScimResourceTypeMeta, ScimSchema,
+    ScimSchemaAttribute, ScimSchemaExtension, ScimSchemaListResponse, ScimSchemaMeta,
+    ScimServiceProviderConfig, ScimServiceProviderConfigBulk, ScimServiceProviderConfigFilter,
+    ScimServiceProviderConfigMeta, ScimServiceProviderConfigPatch, ScimUser, ScimUserAddress,
+    ScimUserEmail, ScimUserEntitlement, ScimUserGroup, ScimUserIm, ScimUserMeta, ScimUserName,
+    ScimUserPhoneNumber, ScimUserPhoto, ScimUserPostRequest, ScimUserPutRequest, ScimUserRole,
+    ScimX509Certificate,
+};
+pub use shared::{
+    ApiResponse, AssignedRole, AssignedRoleRoletype, IpAccessListEntry, IpAccessListEntryResponse,
+    IpAccessListPatch, License, ResourceTagsV1, ResourceTagsV1Response,
+};
+pub use udfs::{
+    Pagination, Udf, UdfArgument, UdfArgumentResponse, UdfAttachment, UdfAttachmentListResponse,
+    UdfAttachmentStatus, UdfCreateRequest, UdfCreateRequestV1, UdfCreateRequestV1Type,
+    UdfCreateRequestV2, UdfCreateRequestV2Type, UdfListResponse, UdfRuntime, UdfSandboxType,
+    UdfSandboxVersion, UdfStatus, UdfType, UdfUploadSession, UdfVersionCreateRequest,
+    UdfVersionCreateRequestV1, UdfVersionCreateRequestV1Type, UdfVersionCreateRequestV2,
+    UdfVersionCreateRequestV2Type, UdfVersionListResponse,
+};
 
 /// `pgHaType` enum from the ClickHouse Cloud API.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
@@ -517,377 +582,6 @@ impl PgVersion {
     pub const VALUES: &'static [&'static str] = &["18", "17"];
 }
 
-/// Inline enum for `Activity.actorType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ActivityActortype {
-    #[serde(rename = "user")]
-    #[default]
-    User,
-    #[serde(rename = "support")]
-    Support,
-    #[serde(rename = "system")]
-    System,
-    #[serde(rename = "api")]
-    Api,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ActivityActortype {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::User => write!(f, "user"),
-            Self::Support => write!(f, "support"),
-            Self::System => write!(f, "system"),
-            Self::Api => write!(f, "api"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Activity.keyUpdateType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ActivityKeyupdatetype {
-    #[serde(rename = "created")]
-    #[default]
-    Created,
-    #[serde(rename = "deleted")]
-    Deleted,
-    #[serde(rename = "name-changed")]
-    Name_changed,
-    #[serde(rename = "role-changed")]
-    Role_changed,
-    #[serde(rename = "state-changed")]
-    State_changed,
-    #[serde(rename = "date-changed")]
-    Date_changed,
-    #[serde(rename = "ip-access-list-changed")]
-    Ip_access_list_changed,
-    #[serde(rename = "org-role-changed")]
-    Org_role_changed,
-    #[serde(rename = "default-service-role-changed")]
-    Default_service_role_changed,
-    #[serde(rename = "service-role-changed")]
-    Service_role_changed,
-    #[serde(rename = "roles-v2-changed")]
-    Roles_v2_changed,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ActivityKeyupdatetype {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Created => write!(f, "created"),
-            Self::Deleted => write!(f, "deleted"),
-            Self::Name_changed => write!(f, "name-changed"),
-            Self::Role_changed => write!(f, "role-changed"),
-            Self::State_changed => write!(f, "state-changed"),
-            Self::Date_changed => write!(f, "date-changed"),
-            Self::Ip_access_list_changed => write!(f, "ip-access-list-changed"),
-            Self::Org_role_changed => write!(f, "org-role-changed"),
-            Self::Default_service_role_changed => write!(f, "default-service-role-changed"),
-            Self::Service_role_changed => write!(f, "service-role-changed"),
-            Self::Roles_v2_changed => write!(f, "roles-v2-changed"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Activity.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ActivityType {
-    #[serde(rename = "create_organization")]
-    #[default]
-    Create_organization,
-    #[serde(rename = "organization_update_name")]
-    Organization_update_name,
-    #[serde(rename = "transfer_service_in")]
-    Transfer_service_in,
-    #[serde(rename = "transfer_service_out")]
-    Transfer_service_out,
-    #[serde(rename = "save_payment_method")]
-    Save_payment_method,
-    #[serde(rename = "marketplace_subscription")]
-    Marketplace_subscription,
-    #[serde(rename = "migrate_marketplace_billing_details_in")]
-    Migrate_marketplace_billing_details_in,
-    #[serde(rename = "migrate_marketplace_billing_details_out")]
-    Migrate_marketplace_billing_details_out,
-    #[serde(rename = "organization_update_tier")]
-    Organization_update_tier,
-    #[serde(rename = "organization_invite_create")]
-    Organization_invite_create,
-    #[serde(rename = "organization_invite_delete")]
-    Organization_invite_delete,
-    #[serde(rename = "organization_member_join")]
-    Organization_member_join,
-    #[serde(rename = "organization_member_add")]
-    Organization_member_add,
-    #[serde(rename = "organization_member_leave")]
-    Organization_member_leave,
-    #[serde(rename = "organization_member_delete")]
-    Organization_member_delete,
-    #[serde(rename = "organization_member_update_role")]
-    Organization_member_update_role,
-    #[serde(rename = "organization_member_update_roles")]
-    Organization_member_update_roles,
-    #[serde(rename = "organization_member_update_mfa_method")]
-    Organization_member_update_mfa_method,
-    #[serde(rename = "organization_saml_connection_create")]
-    Organization_saml_connection_create,
-    #[serde(rename = "organization_saml_connection_update")]
-    Organization_saml_connection_update,
-    #[serde(rename = "user_login")]
-    User_login,
-    #[serde(rename = "user_login_failed")]
-    User_login_failed,
-    #[serde(rename = "user_logout")]
-    User_logout,
-    #[serde(rename = "key_create")]
-    Key_create,
-    #[serde(rename = "key_delete")]
-    Key_delete,
-    #[serde(rename = "openapi_key_update")]
-    Openapi_key_update,
-    #[serde(rename = "service_create")]
-    Service_create,
-    #[serde(rename = "service_start")]
-    Service_start,
-    #[serde(rename = "service_stop")]
-    Service_stop,
-    #[serde(rename = "service_awaken")]
-    Service_awaken,
-    #[serde(rename = "service_idle")]
-    Service_idle,
-    #[serde(rename = "service_running")]
-    Service_running,
-    #[serde(rename = "service_partially_running")]
-    Service_partially_running,
-    #[serde(rename = "service_delete")]
-    Service_delete,
-    #[serde(rename = "service_update_name")]
-    Service_update_name,
-    #[serde(rename = "service_update_ip_access_list")]
-    Service_update_ip_access_list,
-    #[serde(rename = "service_update_autoscaling_memory")]
-    Service_update_autoscaling_memory,
-    #[serde(rename = "service_update_autoscaling_idling")]
-    Service_update_autoscaling_idling,
-    #[serde(rename = "service_update_password")]
-    Service_update_password,
-    #[serde(rename = "service_update_autoscaling_replicas")]
-    Service_update_autoscaling_replicas,
-    #[serde(rename = "service_update_max_allowable_replicas")]
-    Service_update_max_allowable_replicas,
-    #[serde(rename = "service_update_backup_configuration")]
-    Service_update_backup_configuration,
-    #[serde(rename = "service_update_snapshot_configuration")]
-    Service_update_snapshot_configuration,
-    #[serde(rename = "service_restore_backup")]
-    Service_restore_backup,
-    #[serde(rename = "service_update_release_channel")]
-    Service_update_release_channel,
-    #[serde(rename = "service_update_gpt_usage_consent")]
-    Service_update_gpt_usage_consent,
-    #[serde(rename = "service_update_private_endpoints")]
-    Service_update_private_endpoints,
-    #[serde(rename = "service_import_to_organization")]
-    Service_import_to_organization,
-    #[serde(rename = "service_export_from_organization")]
-    Service_export_from_organization,
-    #[serde(rename = "service_maintenance_start")]
-    Service_maintenance_start,
-    #[serde(rename = "service_maintenance_end")]
-    Service_maintenance_end,
-    #[serde(rename = "service_update_core_dump")]
-    Service_update_core_dump,
-    #[serde(rename = "backup_delete")]
-    Backup_delete,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ActivityType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Create_organization => write!(f, "create_organization"),
-            Self::Organization_update_name => write!(f, "organization_update_name"),
-            Self::Transfer_service_in => write!(f, "transfer_service_in"),
-            Self::Transfer_service_out => write!(f, "transfer_service_out"),
-            Self::Save_payment_method => write!(f, "save_payment_method"),
-            Self::Marketplace_subscription => write!(f, "marketplace_subscription"),
-            Self::Migrate_marketplace_billing_details_in => {
-                write!(f, "migrate_marketplace_billing_details_in")
-            }
-            Self::Migrate_marketplace_billing_details_out => {
-                write!(f, "migrate_marketplace_billing_details_out")
-            }
-            Self::Organization_update_tier => write!(f, "organization_update_tier"),
-            Self::Organization_invite_create => write!(f, "organization_invite_create"),
-            Self::Organization_invite_delete => write!(f, "organization_invite_delete"),
-            Self::Organization_member_join => write!(f, "organization_member_join"),
-            Self::Organization_member_add => write!(f, "organization_member_add"),
-            Self::Organization_member_leave => write!(f, "organization_member_leave"),
-            Self::Organization_member_delete => write!(f, "organization_member_delete"),
-            Self::Organization_member_update_role => write!(f, "organization_member_update_role"),
-            Self::Organization_member_update_roles => {
-                write!(f, "organization_member_update_roles")
-            }
-            Self::Organization_member_update_mfa_method => {
-                write!(f, "organization_member_update_mfa_method")
-            }
-            Self::Organization_saml_connection_create => {
-                write!(f, "organization_saml_connection_create")
-            }
-            Self::Organization_saml_connection_update => {
-                write!(f, "organization_saml_connection_update")
-            }
-            Self::User_login => write!(f, "user_login"),
-            Self::User_login_failed => write!(f, "user_login_failed"),
-            Self::User_logout => write!(f, "user_logout"),
-            Self::Key_create => write!(f, "key_create"),
-            Self::Key_delete => write!(f, "key_delete"),
-            Self::Openapi_key_update => write!(f, "openapi_key_update"),
-            Self::Service_create => write!(f, "service_create"),
-            Self::Service_start => write!(f, "service_start"),
-            Self::Service_stop => write!(f, "service_stop"),
-            Self::Service_awaken => write!(f, "service_awaken"),
-            Self::Service_idle => write!(f, "service_idle"),
-            Self::Service_running => write!(f, "service_running"),
-            Self::Service_partially_running => write!(f, "service_partially_running"),
-            Self::Service_delete => write!(f, "service_delete"),
-            Self::Service_update_name => write!(f, "service_update_name"),
-            Self::Service_update_ip_access_list => write!(f, "service_update_ip_access_list"),
-            Self::Service_update_autoscaling_memory => {
-                write!(f, "service_update_autoscaling_memory")
-            }
-            Self::Service_update_autoscaling_idling => {
-                write!(f, "service_update_autoscaling_idling")
-            }
-            Self::Service_update_password => write!(f, "service_update_password"),
-            Self::Service_update_autoscaling_replicas => {
-                write!(f, "service_update_autoscaling_replicas")
-            }
-            Self::Service_update_max_allowable_replicas => {
-                write!(f, "service_update_max_allowable_replicas")
-            }
-            Self::Service_update_backup_configuration => {
-                write!(f, "service_update_backup_configuration")
-            }
-            Self::Service_update_snapshot_configuration => {
-                write!(f, "service_update_snapshot_configuration")
-            }
-            Self::Service_restore_backup => write!(f, "service_restore_backup"),
-            Self::Service_update_release_channel => write!(f, "service_update_release_channel"),
-            Self::Service_update_gpt_usage_consent => write!(f, "service_update_gpt_usage_consent"),
-            Self::Service_update_private_endpoints => write!(f, "service_update_private_endpoints"),
-            Self::Service_import_to_organization => write!(f, "service_import_to_organization"),
-            Self::Service_export_from_organization => write!(f, "service_export_from_organization"),
-            Self::Service_maintenance_start => write!(f, "service_maintenance_start"),
-            Self::Service_maintenance_end => write!(f, "service_maintenance_end"),
-            Self::Service_update_core_dump => write!(f, "service_update_core_dump"),
-            Self::Backup_delete => write!(f, "backup_delete"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ApiKey.state`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ApiKeyState {
-    #[serde(rename = "enabled")]
-    #[default]
-    Enabled,
-    #[serde(rename = "disabled")]
-    Disabled,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ApiKeyState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Enabled => write!(f, "enabled"),
-            Self::Disabled => write!(f, "disabled"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ApiKeyPatchRequest.state`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ApiKeyPatchRequestState {
-    #[serde(rename = "enabled")]
-    #[default]
-    Enabled,
-    #[serde(rename = "disabled")]
-    Disabled,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ApiKeyPatchRequestState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Enabled => write!(f, "enabled"),
-            Self::Disabled => write!(f, "disabled"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ApiKeyPostRequest.state`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ApiKeyPostRequestState {
-    #[serde(rename = "enabled")]
-    #[default]
-    Enabled,
-    #[serde(rename = "disabled")]
-    Disabled,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ApiKeyPostRequestState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Enabled => write!(f, "enabled"),
-            Self::Disabled => write!(f, "disabled"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AssignedRole.roleType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AssignedRoleRoletype {
-    #[serde(rename = "system")]
-    #[default]
-    System,
-    #[serde(rename = "custom")]
-    Custom,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AssignedRoleRoletype {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::System => write!(f, "system"),
-            Self::Custom => write!(f, "custom"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
 /// Inline enum for `AwsBackupBucket.bucketProvider`.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub enum AwsBackupBucketBucketprovider {
@@ -1084,236 +778,6 @@ impl std::fmt::Display for BackupType {
         match self {
             Self::Full => write!(f, "full"),
             Self::Incremental => write!(f, "incremental"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ByocConfig.cloudProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ByocConfigCloudprovider {
-    #[serde(rename = "gcp")]
-    #[default]
-    Gcp,
-    #[serde(rename = "aws")]
-    Aws,
-    #[serde(rename = "azure")]
-    Azure,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ByocConfigCloudprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Gcp => write!(f, "gcp"),
-            Self::Aws => write!(f, "aws"),
-            Self::Azure => write!(f, "azure"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ByocConfig.regionId`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ByocConfigRegionid {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ByocConfigRegionid {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ByocConfig.state`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ByocConfigState {
-    #[serde(rename = "infra-ready")]
-    #[default]
-    Infra_ready,
-    #[serde(rename = "infra-provisioning")]
-    Infra_provisioning,
-    #[serde(rename = "infra-terminated")]
-    Infra_terminated,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ByocConfigState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Infra_ready => write!(f, "infra-ready"),
-            Self::Infra_provisioning => write!(f, "infra-provisioning"),
-            Self::Infra_terminated => write!(f, "infra-terminated"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ByocInfrastructurePostRequest.regionId`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ByocInfrastructurePostRequestRegionid {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ByocInfrastructurePostRequestRegionid {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
             Self::Unknown(s) => write!(f, "{s}"),
         }
     }
@@ -6033,379 +5497,6 @@ impl std::fmt::Display for InstancePrivateEndpointRegion {
     }
 }
 
-/// Inline enum for `Invitation.role`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum InvitationRole {
-    #[serde(rename = "admin")]
-    #[default]
-    Admin,
-    #[serde(rename = "developer")]
-    Developer,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for InvitationRole {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Admin => write!(f, "admin"),
-            Self::Developer => write!(f, "developer"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `InvitationPostRequest.role`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum InvitationPostRequestRole {
-    #[serde(rename = "admin")]
-    #[default]
-    Admin,
-    #[serde(rename = "developer")]
-    Developer,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for InvitationPostRequestRole {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Admin => write!(f, "admin"),
-            Self::Developer => write!(f, "developer"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Member.role`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum MemberRole {
-    #[serde(rename = "admin")]
-    #[default]
-    Admin,
-    #[serde(rename = "developer")]
-    Developer,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for MemberRole {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Admin => write!(f, "admin"),
-            Self::Developer => write!(f, "developer"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `MemberPatchRequest.role`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum MemberPatchRequestRole {
-    #[serde(rename = "admin")]
-    #[default]
-    Admin,
-    #[serde(rename = "developer")]
-    Developer,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for MemberPatchRequestRole {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Admin => write!(f, "admin"),
-            Self::Developer => write!(f, "developer"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `OrganizationPatchPrivateEndpoint.cloudProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum OrganizationPatchPrivateEndpointCloudprovider {
-    #[serde(rename = "gcp")]
-    #[default]
-    Gcp,
-    #[serde(rename = "aws")]
-    Aws,
-    #[serde(rename = "azure")]
-    Azure,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for OrganizationPatchPrivateEndpointCloudprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Gcp => write!(f, "gcp"),
-            Self::Aws => write!(f, "aws"),
-            Self::Azure => write!(f, "azure"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `OrganizationPatchPrivateEndpoint.region`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum OrganizationPatchPrivateEndpointRegion {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for OrganizationPatchPrivateEndpointRegion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `OrganizationPrivateEndpoint.cloudProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum OrganizationPrivateEndpointCloudprovider {
-    #[serde(rename = "gcp")]
-    #[default]
-    Gcp,
-    #[serde(rename = "aws")]
-    Aws,
-    #[serde(rename = "azure")]
-    Azure,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for OrganizationPrivateEndpointCloudprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Gcp => write!(f, "gcp"),
-            Self::Aws => write!(f, "aws"),
-            Self::Azure => write!(f, "azure"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `OrganizationPrivateEndpoint.region`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum OrganizationPrivateEndpointRegion {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for OrganizationPrivateEndpointRegion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `OrganizationQuota.quotaCode`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum OrganizationQuotaQuotacode {
-    #[serde(rename = "services-per-organization")]
-    #[default]
-    Services_per_organization,
-    #[serde(rename = "postgres-services-per-organization")]
-    Postgres_services_per_organization,
-    #[serde(rename = "replicas-per-warehouse")]
-    Replicas_per_warehouse,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for OrganizationQuotaQuotacode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Services_per_organization => write!(f, "services-per-organization"),
-            Self::Postgres_services_per_organization => {
-                write!(f, "postgres-services-per-organization")
-            }
-            Self::Replicas_per_warehouse => write!(f, "replicas-per-warehouse"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `OrganizationQuota.scope`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum OrganizationQuotaScope {
-    #[serde(rename = "organization")]
-    #[default]
-    Organization,
-    #[serde(rename = "warehouse")]
-    Warehouse,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for OrganizationQuotaScope {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Organization => write!(f, "organization"),
-            Self::Warehouse => write!(f, "warehouse"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
 /// Inline enum for `PostgresServiceSetState.command`.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub enum PostgresServiceSetStateCommand {
@@ -6427,94 +5518,6 @@ impl std::fmt::Display for PostgresServiceSetStateCommand {
             Self::Restart => write!(f, "restart"),
             Self::Promote => write!(f, "promote"),
             Self::Switchover => write!(f, "switchover"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `RBACPolicy.allowDeny`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum RBACPolicyAllowdeny {
-    #[default]
-    ALLOW,
-    DENY,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for RBACPolicyAllowdeny {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ALLOW => write!(f, "ALLOW"),
-            Self::DENY => write!(f, "DENY"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `RBACPolicyCreateRequest.allowDeny`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum RBACPolicyCreateRequestAllowdeny {
-    #[default]
-    ALLOW,
-    DENY,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for RBACPolicyCreateRequestAllowdeny {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ALLOW => write!(f, "ALLOW"),
-            Self::DENY => write!(f, "DENY"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `RBACPolicyTags.roleV2`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum RBACPolicyTagsRolev2 {
-    #[serde(rename = "sql-console-readonly")]
-    #[default]
-    Sql_console_readonly,
-    #[serde(rename = "sql-console-admin")]
-    Sql_console_admin,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for RBACPolicyTagsRolev2 {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Sql_console_readonly => write!(f, "sql-console-readonly"),
-            Self::Sql_console_admin => write!(f, "sql-console-admin"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `RBACRole.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum RBACRoleType {
-    #[serde(rename = "system")]
-    #[default]
-    System,
-    #[serde(rename = "custom")]
-    Custom,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for RBACRoleType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::System => write!(f, "system"),
-            Self::Custom => write!(f, "custom"),
             Self::Unknown(s) => write!(f, "{s}"),
         }
     }
@@ -6594,32 +5597,6 @@ impl std::fmt::Display for ReversePrivateEndpointType {
             Self::VPC_RESOURCE => write!(f, "VPC_RESOURCE"),
             Self::MSK_MULTI_VPC => write!(f, "MSK_MULTI_VPC"),
             Self::GCP_PSC_SERVICE_ATTACHMENT => write!(f, "GCP_PSC_SERVICE_ATTACHMENT"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ScimPatchOperation.op`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ScimPatchOperationOp {
-    #[serde(rename = "add")]
-    #[default]
-    Add,
-    #[serde(rename = "replace")]
-    Replace,
-    #[serde(rename = "remove")]
-    Remove,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ScimPatchOperationOp {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Add => write!(f, "add"),
-            Self::Replace => write!(f, "replace"),
-            Self::Remove => write!(f, "remove"),
             Self::Unknown(s) => write!(f, "{s}"),
         }
     }
@@ -9189,130 +8166,6 @@ pub type PgTags = Vec<ResourceTagsV1>;
 /// [`ResourceTagsV1Response`].
 pub type PgTagsResponse = Vec<ResourceTagsV1Response>;
 
-/// `Activity` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Activity {
-    #[serde(rename = "actorDetails", skip_serializing_if = "Option::is_none")]
-    pub actor_details: Option<String>,
-    #[serde(rename = "actorId", skip_serializing_if = "Option::is_none")]
-    pub actor_id: Option<String>,
-    #[serde(rename = "actorIpAddress", skip_serializing_if = "Option::is_none")]
-    pub actor_ip_address: Option<String>,
-    #[serde(rename = "actorType", skip_serializing_if = "Option::is_none")]
-    pub actor_type: Option<ActivityActortype>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(rename = "keyUpdateType", skip_serializing_if = "Option::is_none")]
-    pub key_update_type: Option<ActivityKeyupdatetype>,
-    #[serde(rename = "organizationId", skip_serializing_if = "Option::is_none")]
-    pub organization_id: Option<String>,
-    #[serde(rename = "serviceId", skip_serializing_if = "Option::is_none")]
-    pub service_id: Option<String>,
-    #[serde(rename = "targetKeyId", skip_serializing_if = "Option::is_none")]
-    pub target_key_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<ActivityType>,
-    #[serde(rename = "userAgent", skip_serializing_if = "Option::is_none")]
-    pub user_agent: Option<String>,
-}
-
-/// `ApiKey` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ApiKey {
-    #[serde(rename = "assignedRoles", skip_serializing_if = "Option::is_none")]
-    pub assigned_roles: Option<Vec<AssignedRole>>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(rename = "expireAt", skip_serializing_if = "Option::is_none")]
-    pub expire_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-    #[serde(rename = "ipAccessList", skip_serializing_if = "Option::is_none")]
-    pub ip_access_list: Option<Vec<IpAccessListEntryResponse>>,
-    #[serde(rename = "keySuffix", skip_serializing_if = "Option::is_none")]
-    pub key_suffix: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<ApiKeyState>,
-    #[serde(rename = "usedAt", skip_serializing_if = "Option::is_none")]
-    pub used_at: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-/// `ApiKeyHashData` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ApiKeyHashData {
-    #[serde(rename = "keyIdHash")]
-    pub key_id_hash: String,
-    #[serde(rename = "keyIdSuffix")]
-    pub key_id_suffix: String,
-    #[serde(rename = "keySecretHash")]
-    pub key_secret_hash: String,
-}
-
-/// `ApiKeyPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ApiKeyPatchRequest {
-    #[serde(rename = "assignedRoleIds", skip_serializing_if = "Option::is_none")]
-    pub assigned_role_ids: Option<Vec<uuid::Uuid>>,
-    #[serde(rename = "expireAt", skip_serializing_if = "Option::is_none")]
-    pub expire_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(rename = "ipAccessList", skip_serializing_if = "Option::is_none")]
-    pub ip_access_list: Option<Vec<IpAccessListEntry>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<ApiKeyPatchRequestState>,
-}
-
-/// `ApiKeyPostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ApiKeyPostRequest {
-    #[serde(rename = "assignedRoleIds")]
-    pub assigned_role_ids: Vec<uuid::Uuid>,
-    #[serde(rename = "expireAt", skip_serializing_if = "Option::is_none")]
-    pub expire_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(rename = "hashData", skip_serializing_if = "Option::is_none")]
-    pub hash_data: Option<ApiKeyHashData>,
-    #[serde(rename = "ipAccessList")]
-    pub ip_access_list: Vec<IpAccessListEntry>,
-    pub name: String,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<String>>,
-    pub state: ApiKeyPostRequestState,
-}
-
-/// `ApiKeyPostResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ApiKeyPostResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub key: Option<ApiKey>,
-    #[serde(rename = "keyId", skip_serializing_if = "Option::is_none")]
-    pub key_id: Option<String>,
-    #[serde(rename = "keySecret", skip_serializing_if = "Option::is_none")]
-    pub key_secret: Option<String>,
-}
-
-/// `AssignedRole` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AssignedRole {
-    #[serde(rename = "roleId", skip_serializing_if = "Option::is_none")]
-    pub role_id: Option<uuid::Uuid>,
-    #[serde(rename = "roleName", skip_serializing_if = "Option::is_none")]
-    pub role_name: Option<String>,
-    #[serde(rename = "roleType", skip_serializing_if = "Option::is_none")]
-    pub role_type: Option<AssignedRoleRoletype>,
-}
-
 /// `AwsBackupBucket` from the ClickHouse Cloud API.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub struct AwsBackupBucket {
@@ -9487,45 +8340,6 @@ pub struct BasePostgresService {
     pub region: PgRegion,
     pub size: PgSize,
     pub tags: PgTags,
-}
-
-/// `ByocConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ByocConfig {
-    #[serde(rename = "accountName", skip_serializing_if = "Option::is_none")]
-    pub account_name: Option<String>,
-    #[serde(rename = "cloudProvider", skip_serializing_if = "Option::is_none")]
-    pub cloud_provider: Option<ByocConfigCloudprovider>,
-    #[serde(rename = "displayName", skip_serializing_if = "Option::is_none")]
-    pub display_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(rename = "regionId", skip_serializing_if = "Option::is_none")]
-    pub region_id: Option<ByocConfigRegionid>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<ByocConfigState>,
-}
-
-/// `ByocInfrastructurePatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ByocInfrastructurePatchRequest {
-    #[serde(rename = "displayName", skip_serializing_if = "Option::is_none")]
-    pub display_name: Option<String>,
-}
-
-/// `ByocInfrastructurePostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ByocInfrastructurePostRequest {
-    #[serde(rename = "accountId")]
-    pub account_id: String,
-    #[serde(rename = "availabilityZoneSuffixes")]
-    pub availability_zone_suffixes: Vec<String>,
-    #[serde(rename = "displayName")]
-    pub display_name: String,
-    #[serde(rename = "regionId")]
-    pub region_id: ByocInfrastructurePostRequestRegionid,
-    #[serde(rename = "vpcCidrRange")]
-    pub vpc_cidr_range: String,
 }
 
 /// `ClickPipe` from the ClickHouse Cloud API.
@@ -14288,102 +13102,6 @@ pub struct InstanceTagsPatch {
     pub remove: Vec<ResourceTagsV1>,
 }
 
-/// `Invitation` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Invitation {
-    #[serde(rename = "assignedRoles", skip_serializing_if = "Option::is_none")]
-    pub assigned_roles: Option<Vec<AssignedRole>>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub email: Option<String>,
-    #[serde(rename = "expireAt", skip_serializing_if = "Option::is_none")]
-    pub expire_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub role: Option<InvitationRole>,
-}
-
-/// `InvitationPostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct InvitationPostRequest {
-    #[serde(rename = "assignedRoleIds")]
-    pub assigned_role_ids: Vec<String>,
-    pub email: String,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub role: Option<InvitationPostRequestRole>,
-}
-
-/// `IpAccessListEntry` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct IpAccessListEntry {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    pub source: String,
-}
-
-/// `IpAccessListEntry` from the ClickHouse Cloud API, in response position.
-///
-/// Response variant of [`IpAccessListEntry`]: every field is `Option<T>`, so a
-/// field the API drops or sends as `null` deserializes to `None` instead of
-/// failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct IpAccessListEntryResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-}
-
-/// `IpAccessListPatch` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct IpAccessListPatch {
-    pub add: Vec<IpAccessListEntry>,
-    pub remove: Vec<IpAccessListEntry>,
-}
-
-/// `License` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct License {
-    #[serde(rename = "environmentFingerprint")]
-    pub environment_fingerprint: String,
-    pub expiration: String,
-    pub id: String,
-    pub memory: String,
-    pub name: String,
-}
-
-/// `Member` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Member {
-    #[serde(rename = "assignedRoles", skip_serializing_if = "Option::is_none")]
-    pub assigned_roles: Option<Vec<AssignedRole>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub email: Option<String>,
-    #[serde(rename = "joinedAt", skip_serializing_if = "Option::is_none")]
-    pub joined_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub role: Option<MemberRole>,
-    #[serde(rename = "userId", skip_serializing_if = "Option::is_none")]
-    pub user_id: Option<String>,
-}
-
-/// `MemberPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct MemberPatchRequest {
-    #[serde(rename = "assignedRoleIds", skip_serializing_if = "Option::is_none")]
-    pub assigned_role_ids: Option<Vec<String>>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub role: Option<MemberPatchRequestRole>,
-}
-
 /// `MskIamUser` from the ClickHouse Cloud API.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub struct MskIamUser {
@@ -14399,93 +13117,6 @@ pub struct MutualTLS {
     pub certificate: String,
     #[serde(rename = "privateKey")]
     pub private_key: String,
-}
-
-/// `Organization` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Organization {
-    #[serde(rename = "byocConfig", skip_serializing_if = "Option::is_none")]
-    pub byoc_config: Option<Vec<ByocConfig>>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(rename = "enableCoreDumps", skip_serializing_if = "Option::is_none")]
-    pub enable_core_dumps: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "privateEndpoints", skip_serializing_if = "Option::is_none")]
-    pub private_endpoints: Option<Vec<OrganizationPrivateEndpoint>>,
-}
-
-/// `OrganizationCloudRegionPrivateEndpointConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct OrganizationCloudRegionPrivateEndpointConfig {
-    #[serde(rename = "endpointServiceId", skip_serializing_if = "Option::is_none")]
-    pub endpoint_service_id: Option<String>,
-}
-
-/// `OrganizationPatchPrivateEndpoint` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct OrganizationPatchPrivateEndpoint {
-    #[serde(rename = "cloudProvider")]
-    pub cloud_provider: OrganizationPatchPrivateEndpointCloudprovider,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    pub id: String,
-    pub region: OrganizationPatchPrivateEndpointRegion,
-}
-
-/// `OrganizationPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct OrganizationPatchRequest {
-    #[serde(rename = "enableCoreDumps", skip_serializing_if = "Option::is_none")]
-    pub enable_core_dumps: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "privateEndpoints", skip_serializing_if = "Option::is_none")]
-    pub private_endpoints: Option<OrganizationPrivateEndpointsPatch>,
-}
-
-/// `OrganizationPrivateEndpoint` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct OrganizationPrivateEndpoint {
-    #[serde(rename = "cloudProvider", skip_serializing_if = "Option::is_none")]
-    pub cloud_provider: Option<OrganizationPrivateEndpointCloudprovider>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub region: Option<OrganizationPrivateEndpointRegion>,
-}
-
-/// `OrganizationPrivateEndpointsPatch` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct OrganizationPrivateEndpointsPatch {
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub add: Option<Vec<OrganizationPatchPrivateEndpoint>>,
-    pub remove: Vec<OrganizationPatchPrivateEndpoint>,
-}
-
-/// `OrganizationQuota` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct OrganizationQuota {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub adjustable: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "quotaCode", skip_serializing_if = "Option::is_none")]
-    pub quota_code: Option<OrganizationQuotaQuotacode>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub scope: Option<OrganizationQuotaScope>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub usage: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value: Option<i64>,
 }
 
 /// `PLAIN` from the ClickHouse Cloud API.
@@ -14849,104 +13480,6 @@ pub struct PrivateEndpointConfig {
     pub private_dns_hostname: Option<String>,
 }
 
-/// `RBACPolicy` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct RBACPolicy {
-    #[serde(rename = "allowDeny", skip_serializing_if = "Option::is_none")]
-    pub allow_deny: Option<RBACPolicyAllowdeny>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub permissions: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub resources: Option<Vec<String>>,
-    #[serde(rename = "roleId", skip_serializing_if = "Option::is_none")]
-    pub role_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<RBACPolicyTagsResponse>,
-    #[serde(rename = "tenantId", skip_serializing_if = "Option::is_none")]
-    pub tenant_id: Option<String>,
-}
-
-/// `RBACPolicyCreateRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct RBACPolicyCreateRequest {
-    #[serde(rename = "allowDeny")]
-    pub allow_deny: RBACPolicyCreateRequestAllowdeny,
-    pub permissions: Vec<String>,
-    pub resources: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<RBACPolicyTags>,
-}
-
-/// `RBACPolicyTags` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct RBACPolicyTags {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub grants: Option<Vec<String>>,
-    #[serde(rename = "roleV2", skip_serializing_if = "Option::is_none")]
-    pub role_v2: Option<RBACPolicyTagsRolev2>,
-}
-
-/// `RBACPolicyTags` from the ClickHouse Cloud API, in response position.
-///
-/// Response variant of [`RBACPolicyTags`]: every field is `Option<T>`, so a
-/// field the API drops or sends as `null` deserializes to `None` instead of
-/// failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct RBACPolicyTagsResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub grants: Option<Vec<String>>,
-    #[serde(rename = "roleV2", skip_serializing_if = "Option::is_none")]
-    pub role_v2: Option<RBACPolicyTagsRolev2>,
-}
-
-/// `RBACRole` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct RBACRole {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub actors: Option<Vec<String>>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "ownerId", skip_serializing_if = "Option::is_none")]
-    pub owner_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub policies: Option<Vec<RBACPolicy>>,
-    #[serde(rename = "tenantId", skip_serializing_if = "Option::is_none")]
-    pub tenant_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<RBACRoleType>,
-    #[serde(rename = "updatedAt", skip_serializing_if = "Option::is_none")]
-    pub updated_at: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-/// `ResourceTagsV1` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ResourceTagsV1 {
-    pub key: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value: Option<String>,
-}
-
-/// `ResourceTagsV1` from the ClickHouse Cloud API, in response position.
-///
-/// Response variant of [`ResourceTagsV1`]: every field is `Option<T>`, so a
-/// field the API drops or sends as `null` deserializes to `None` instead of
-/// failing. Writing a fetched tag back to the API goes through
-/// `TryFrom<ResourceTagsV1Response>` (see [`crate::convert`]), because a tag
-/// without a key cannot be sent.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ResourceTagsV1Response {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub key: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value: Option<String>,
-}
-
 /// `ReversePrivateEndpoint` from the ClickHouse Cloud API.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub struct ReversePrivateEndpoint {
@@ -14995,22 +13528,6 @@ pub struct ReversePrivateEndpoint {
         skip_serializing_if = "Option::is_none"
     )]
     pub vpc_resource_share_arn: Option<String>,
-}
-
-/// `RoleCreateRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct RoleCreateRequest {
-    pub actors: Vec<String>,
-    pub name: String,
-    pub policies: Vec<RBACPolicyCreateRequest>,
-}
-
-/// `RoleUpdateRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct RoleUpdateRequest {
-    pub actors: Vec<String>,
-    pub name: String,
-    pub policies: Vec<RBACPolicyCreateRequest>,
 }
 
 /// `ScalingSchedule` from the ClickHouse Cloud API.
@@ -15105,589 +13622,6 @@ pub struct ScalingScheduleEntryRequest {
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub struct ScalingSchedulePostRequest {
     pub entries: Vec<ScalingScheduleEntryRequest>,
-}
-
-// The `Scim*` family below is strict in both directions, deliberately. The spec
-// defines the SCIM schemas but declares no SCIM path, so no `Client` method
-// sends or returns one: the family is reachable from neither a request root nor
-// a response root, and the analyzer resolves such operation-unreferenced schemas
-// in request position. Making the SCIM list/response envelopes all-`Option`
-// would therefore report `FieldOptionalityMismatch` drift while protecting no
-// actual response. `scim_models_are_outside_the_response_tree` in
-// `tests/spec_coverage_test.rs` pins that premise: if SCIM operations are ever
-// added to `client.rs`, the envelopes they return must be split first.
-
-/// `ScimEnterpriseManager` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimEnterpriseManager {
-    #[serde(rename = "displayName")]
-    pub display_name: String,
-    pub value: String,
-}
-
-/// `ScimEnterpriseUser` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimEnterpriseUser {
-    #[serde(rename = "costCenter")]
-    pub cost_center: String,
-    pub department: String,
-    pub division: String,
-    #[serde(rename = "employeeNumber")]
-    pub employee_number: String,
-    pub manager: ScimEnterpriseManager,
-    pub organization: String,
-}
-
-/// `ScimGroup` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimGroup {
-    #[serde(rename = "displayName")]
-    pub display_name: String,
-    #[serde(rename = "externalId", skip_serializing_if = "Option::is_none")]
-    pub external_id: Option<String>,
-    pub id: uuid::Uuid,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub members: Option<Vec<ScimGroupMember>>,
-    pub meta: ScimGroupMeta,
-    pub schemas: Vec<String>,
-}
-
-/// `ScimGroupListResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimGroupListResponse {
-    #[serde(rename = "Resources")]
-    pub resources: Vec<ScimGroup>,
-    #[serde(rename = "itemsPerPage")]
-    pub items_per_page: i64,
-    pub schemas: Vec<String>,
-    #[serde(rename = "startIndex")]
-    pub start_index: i64,
-    #[serde(rename = "totalResults")]
-    pub total_results: i64,
-}
-
-/// `ScimGroupMember` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimGroupMember {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub display: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<String>,
-    pub value: String,
-}
-
-/// `ScimGroupMeta` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimGroupMeta {
-    pub created: chrono::DateTime<chrono::Utc>,
-    #[serde(rename = "lastModified")]
-    pub last_modified: chrono::DateTime<chrono::Utc>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub location: Option<String>,
-    #[serde(rename = "resourceType")]
-    pub resource_type: String,
-}
-
-/// `ScimGroupPostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimGroupPostRequest {
-    #[serde(rename = "displayName")]
-    pub display_name: String,
-    #[serde(rename = "externalId", skip_serializing_if = "Option::is_none")]
-    pub external_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub members: Option<Vec<ScimGroupMember>>,
-    pub schemas: Vec<String>,
-}
-
-/// `ScimGroupPutRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimGroupPutRequest {
-    #[serde(rename = "displayName")]
-    pub display_name: String,
-    #[serde(rename = "externalId", skip_serializing_if = "Option::is_none")]
-    pub external_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub members: Option<Vec<ScimGroupMember>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub meta: Option<ScimGroupMeta>,
-    pub schemas: Vec<String>,
-}
-
-/// `ScimListResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimListResponse {
-    #[serde(rename = "Resources")]
-    pub resources: Vec<ScimUser>,
-    #[serde(rename = "itemsPerPage")]
-    pub items_per_page: i64,
-    pub schemas: Vec<String>,
-    #[serde(rename = "startIndex")]
-    pub start_index: i64,
-    #[serde(rename = "totalResults")]
-    pub total_results: i64,
-}
-
-/// `ScimPatchOp` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimPatchOp {
-    #[serde(rename = "Operations")]
-    pub operations: Vec<ScimPatchOperation>,
-    pub schemas: Vec<String>,
-}
-
-/// `ScimPatchOperation` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimPatchOperation {
-    pub op: ScimPatchOperationOp,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value: Option<String>,
-}
-
-/// `ScimUser` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUser {
-    pub active: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub addresses: Option<Vec<ScimUserAddress>>,
-    #[serde(rename = "displayName", skip_serializing_if = "Option::is_none")]
-    pub display_name: Option<String>,
-    pub emails: Vec<ScimUserEmail>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub entitlements: Option<Vec<ScimUserEntitlement>>,
-    #[serde(
-        rename = "urn:ietf:params:scim:schemas:extension:enterprise:2.0:User",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub enterprise_user: Option<ScimEnterpriseUser>,
-    #[serde(rename = "externalId", skip_serializing_if = "Option::is_none")]
-    pub external_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub groups: Option<Vec<ScimUserGroup>>,
-    pub id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ims: Option<Vec<ScimUserIm>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub locale: Option<String>,
-    pub meta: ScimUserMeta,
-    pub name: ScimUserName,
-    #[serde(rename = "nickName", skip_serializing_if = "Option::is_none")]
-    pub nick_name: Option<String>,
-    #[serde(rename = "phoneNumbers", skip_serializing_if = "Option::is_none")]
-    pub phone_numbers: Option<Vec<ScimUserPhoneNumber>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub photos: Option<Vec<ScimUserPhoto>>,
-    #[serde(rename = "preferredLanguage", skip_serializing_if = "Option::is_none")]
-    pub preferred_language: Option<String>,
-    #[serde(rename = "profileUrl", skip_serializing_if = "Option::is_none")]
-    pub profile_url: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<ScimUserRole>>,
-    pub schemas: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timezone: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    #[serde(rename = "userName")]
-    pub user_name: String,
-    #[serde(rename = "userType", skip_serializing_if = "Option::is_none")]
-    pub user_type: Option<String>,
-    #[serde(rename = "x509Certificates", skip_serializing_if = "Option::is_none")]
-    pub x509_certificates: Option<Vec<ScimX509Certificate>>,
-}
-
-/// `ScimUserAddress` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserAddress {
-    pub country: String,
-    pub formatted: String,
-    pub locality: String,
-    #[serde(rename = "postalCode")]
-    pub postal_code: String,
-    pub primary: bool,
-    pub region: String,
-    #[serde(rename = "streetAddress")]
-    pub street_address: String,
-    pub r#type: String,
-}
-
-/// `ScimUserEmail` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserEmail {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub primary: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<String>,
-    pub value: String,
-}
-
-/// `ScimUserEntitlement` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserEntitlement {
-    pub display: String,
-    pub primary: bool,
-    pub r#type: String,
-    pub value: String,
-}
-
-/// `ScimUserGroup` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserGroup {
-    pub display: String,
-    pub r#type: String,
-    pub value: String,
-}
-
-/// `ScimUserIm` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserIm {
-    pub primary: bool,
-    pub r#type: String,
-    pub value: String,
-}
-
-/// `ScimUserMeta` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserMeta {
-    pub created: chrono::DateTime<chrono::Utc>,
-    #[serde(rename = "lastModified")]
-    pub last_modified: chrono::DateTime<chrono::Utc>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub location: Option<String>,
-    #[serde(rename = "resourceType")]
-    pub resource_type: String,
-}
-
-/// `ScimUserName` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserName {
-    #[serde(rename = "familyName")]
-    pub family_name: String,
-    pub formatted: String,
-    #[serde(rename = "givenName")]
-    pub given_name: String,
-    #[serde(rename = "honorificPrefix")]
-    pub honorific_prefix: String,
-    #[serde(rename = "honorificSuffix")]
-    pub honorific_suffix: String,
-    #[serde(rename = "middleName")]
-    pub middle_name: String,
-}
-
-/// `ScimUserPhoneNumber` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserPhoneNumber {
-    pub primary: bool,
-    pub r#type: String,
-    pub value: String,
-}
-
-/// `ScimUserPhoto` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserPhoto {
-    pub primary: bool,
-    pub r#type: String,
-    pub value: String,
-}
-
-/// `ScimUserPostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserPostRequest {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub active: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub addresses: Option<Vec<ScimUserAddress>>,
-    #[serde(rename = "displayName", skip_serializing_if = "Option::is_none")]
-    pub display_name: Option<String>,
-    pub emails: Vec<ScimUserEmail>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub entitlements: Option<Vec<ScimUserEntitlement>>,
-    #[serde(rename = "externalId", skip_serializing_if = "Option::is_none")]
-    pub external_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub groups: Option<Vec<ScimUserGroup>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ims: Option<Vec<ScimUserIm>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub locale: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<ScimUserName>,
-    #[serde(rename = "nickName", skip_serializing_if = "Option::is_none")]
-    pub nick_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub password: Option<String>,
-    #[serde(rename = "phoneNumbers", skip_serializing_if = "Option::is_none")]
-    pub phone_numbers: Option<Vec<ScimUserPhoneNumber>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub photos: Option<Vec<ScimUserPhoto>>,
-    #[serde(rename = "preferredLanguage", skip_serializing_if = "Option::is_none")]
-    pub preferred_language: Option<String>,
-    #[serde(rename = "profileUrl", skip_serializing_if = "Option::is_none")]
-    pub profile_url: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<ScimUserRole>>,
-    pub schemas: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timezone: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    #[serde(
-        rename = "urn:ietf:params:scim:schemas:extension:enterprise:2.0:User",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub urn_ietf_params_scim_schemas_extension_enterprise_2_0_user: Option<ScimEnterpriseUser>,
-    #[serde(rename = "userName")]
-    pub user_name: String,
-    #[serde(rename = "userType", skip_serializing_if = "Option::is_none")]
-    pub user_type: Option<String>,
-    #[serde(rename = "x509Certificates", skip_serializing_if = "Option::is_none")]
-    pub x509_certificates: Option<Vec<ScimX509Certificate>>,
-}
-
-/// `ScimUserPutRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserPutRequest {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub active: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub addresses: Option<Vec<ScimUserAddress>>,
-    #[serde(rename = "displayName", skip_serializing_if = "Option::is_none")]
-    pub display_name: Option<String>,
-    pub emails: Vec<ScimUserEmail>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub entitlements: Option<Vec<ScimUserEntitlement>>,
-    #[serde(rename = "externalId", skip_serializing_if = "Option::is_none")]
-    pub external_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub groups: Option<Vec<ScimUserGroup>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ims: Option<Vec<ScimUserIm>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub locale: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub meta: Option<ScimUserMeta>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<ScimUserName>,
-    #[serde(rename = "nickName", skip_serializing_if = "Option::is_none")]
-    pub nick_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub password: Option<String>,
-    #[serde(rename = "phoneNumbers", skip_serializing_if = "Option::is_none")]
-    pub phone_numbers: Option<Vec<ScimUserPhoneNumber>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub photos: Option<Vec<ScimUserPhoto>>,
-    #[serde(rename = "preferredLanguage", skip_serializing_if = "Option::is_none")]
-    pub preferred_language: Option<String>,
-    #[serde(rename = "profileUrl", skip_serializing_if = "Option::is_none")]
-    pub profile_url: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<ScimUserRole>>,
-    pub schemas: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timezone: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    #[serde(
-        rename = "urn:ietf:params:scim:schemas:extension:enterprise:2.0:User",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub urn_ietf_params_scim_schemas_extension_enterprise_2_0_user: Option<ScimEnterpriseUser>,
-    #[serde(rename = "userName")]
-    pub user_name: String,
-    #[serde(rename = "userType", skip_serializing_if = "Option::is_none")]
-    pub user_type: Option<String>,
-    #[serde(rename = "x509Certificates", skip_serializing_if = "Option::is_none")]
-    pub x509_certificates: Option<Vec<ScimX509Certificate>>,
-}
-
-/// `ScimUserRole` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimUserRole {
-    pub display: String,
-    pub primary: bool,
-    pub r#type: String,
-    pub value: String,
-}
-
-/// `ScimX509Certificate` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimX509Certificate {
-    pub value: String,
-}
-
-/// `ScimAuthenticationScheme` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimAuthenticationScheme {
-    pub description: String,
-    pub name: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub primary: Option<bool>,
-    #[serde(rename = "specUri", skip_serializing_if = "Option::is_none")]
-    pub spec_uri: Option<String>,
-    #[serde(rename = "type")]
-    pub r#type: String,
-}
-
-/// `ScimBooleanFeature` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimBooleanFeature {
-    pub supported: bool,
-}
-
-/// `ScimResourceType` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimResourceType {
-    pub description: String,
-    pub endpoint: String,
-    pub id: String,
-    pub meta: ScimResourceTypeMeta,
-    pub name: String,
-    pub schema: String,
-    #[serde(rename = "schemaExtensions")]
-    pub schema_extensions: Vec<ScimSchemaExtension>,
-    pub schemas: Vec<String>,
-}
-
-/// `ScimResourceTypeListResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimResourceTypeListResponse {
-    #[serde(rename = "Resources")]
-    pub resources: Vec<ScimResourceType>,
-    #[serde(rename = "itemsPerPage")]
-    pub items_per_page: i64,
-    pub schemas: Vec<String>,
-    #[serde(rename = "startIndex")]
-    pub start_index: i64,
-    #[serde(rename = "totalResults")]
-    pub total_results: i64,
-}
-
-/// `ScimResourceTypeMeta` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimResourceTypeMeta {
-    pub location: String,
-    #[serde(rename = "resourceType")]
-    pub resource_type: String,
-}
-
-/// `ScimSchema` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimSchema {
-    pub attributes: Vec<ScimSchemaAttribute>,
-    pub description: String,
-    pub id: String,
-    pub meta: ScimSchemaMeta,
-    pub name: String,
-    pub schemas: Vec<String>,
-}
-
-/// `ScimSchemaAttribute` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimSchemaAttribute {
-    #[serde(rename = "canonicalValues", skip_serializing_if = "Option::is_none")]
-    pub canonical_values: Option<Vec<String>>,
-    #[serde(rename = "caseExact", skip_serializing_if = "Option::is_none")]
-    pub case_exact: Option<bool>,
-    pub description: String,
-    #[serde(rename = "multiValued")]
-    pub multi_valued: bool,
-    pub mutability: String,
-    pub name: String,
-    #[serde(rename = "referenceTypes", skip_serializing_if = "Option::is_none")]
-    pub reference_types: Option<Vec<String>>,
-    pub required: bool,
-    pub returned: String,
-    #[serde(rename = "subAttributes", skip_serializing_if = "Option::is_none")]
-    pub sub_attributes: Option<Vec<ScimSchemaAttribute>>,
-    #[serde(rename = "type")]
-    pub r#type: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uniqueness: Option<String>,
-}
-
-/// `ScimSchemaExtension` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimSchemaExtension {
-    pub required: bool,
-    pub schema: String,
-}
-
-/// `ScimSchemaListResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimSchemaListResponse {
-    #[serde(rename = "Resources")]
-    pub resources: Vec<ScimSchema>,
-    #[serde(rename = "itemsPerPage")]
-    pub items_per_page: i64,
-    pub schemas: Vec<String>,
-    #[serde(rename = "startIndex")]
-    pub start_index: i64,
-    #[serde(rename = "totalResults")]
-    pub total_results: i64,
-}
-
-/// `ScimSchemaMeta` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimSchemaMeta {
-    pub location: String,
-    #[serde(rename = "resourceType")]
-    pub resource_type: String,
-}
-
-/// `ScimServiceProviderConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimServiceProviderConfig {
-    #[serde(rename = "authenticationSchemes")]
-    pub authentication_schemes: Vec<ScimAuthenticationScheme>,
-    pub bulk: ScimServiceProviderConfigBulk,
-    #[serde(rename = "changePassword")]
-    pub change_password: ScimBooleanFeature,
-    #[serde(rename = "documentationUri", skip_serializing_if = "Option::is_none")]
-    pub documentation_uri: Option<String>,
-    pub etag: ScimBooleanFeature,
-    pub filter: ScimServiceProviderConfigFilter,
-    pub meta: ScimServiceProviderConfigMeta,
-    pub patch: ScimServiceProviderConfigPatch,
-    pub schemas: Vec<String>,
-    pub sort: ScimBooleanFeature,
-}
-
-/// `ScimServiceProviderConfigBulk` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimServiceProviderConfigBulk {
-    #[serde(rename = "maxOperations")]
-    pub max_operations: i64,
-    #[serde(rename = "maxPayloadSize")]
-    pub max_payload_size: i64,
-    pub supported: bool,
-}
-
-/// `ScimServiceProviderConfigFilter` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimServiceProviderConfigFilter {
-    #[serde(rename = "maxResults")]
-    pub max_results: i64,
-    pub supported: bool,
-}
-
-/// `ScimServiceProviderConfigMeta` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimServiceProviderConfigMeta {
-    pub location: String,
-    #[serde(rename = "resourceType")]
-    pub resource_type: String,
-}
-
-/// `ScimServiceProviderConfigPatch` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScimServiceProviderConfigPatch {
-    pub supported: bool,
 }
 
 /// `ServicPrivateEndpointePostRequest` from the ClickHouse Cloud API.
@@ -16461,618 +14395,6 @@ pub struct PostgresInstanceUpdateConfigResponse {
     pub pg_bouncer_config: Option<PgBouncerConfigResponse>,
     #[serde(rename = "pgConfig", skip_serializing_if = "Option::is_none")]
     pub pg_config: Option<PgConfigResponse>,
-}
-
-/// `Pagination` from the ClickHouse Cloud API.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Pagination {
-    #[serde(rename = "currentCursor", skip_serializing_if = "Option::is_none")]
-    pub current_cursor: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<i64>,
-    #[serde(rename = "nextCursor", skip_serializing_if = "Option::is_none")]
-    pub next_cursor: Option<String>,
-    #[serde(rename = "totalRecords", skip_serializing_if = "Option::is_none")]
-    pub total_records: Option<i64>,
-}
-
-/// `UdfAttachment` from the ClickHouse Cloud API.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfAttachment {
-    #[serde(rename = "functionName", skip_serializing_if = "Option::is_none")]
-    pub function_name: Option<String>,
-    #[serde(rename = "serviceId", skip_serializing_if = "Option::is_none")]
-    pub service_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<UdfAttachmentStatus>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub version: Option<i64>,
-}
-
-/// Inline enum for `UdfAttachment.status`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfAttachmentStatus {
-    #[serde(rename = "deployed")]
-    #[default]
-    Deployed,
-    #[serde(rename = "deprovisioning")]
-    Deprovisioning,
-    #[serde(rename = "error")]
-    Error,
-    #[serde(rename = "provisioning")]
-    Provisioning,
-    #[serde(rename = "standby")]
-    Standby,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfAttachmentStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Deployed => write!(f, "deployed"),
-            Self::Deprovisioning => write!(f, "deprovisioning"),
-            Self::Error => write!(f, "error"),
-            Self::Provisioning => write!(f, "provisioning"),
-            Self::Standby => write!(f, "standby"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `UdfAttachmentListResponse` from the ClickHouse Cloud API.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfAttachmentListResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub items: Option<Vec<UdfAttachment>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pagination: Option<Pagination>,
-}
-
-/// A UDF argument sent in a create or version-create request.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfArgument {
-    pub name: String,
-    pub r#type: String,
-}
-
-/// `Udf` from the ClickHouse Cloud API.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Udf {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub arguments: Option<Vec<UdfArgumentResponse>>,
-    #[serde(rename = "commandReadTimeout", skip_serializing_if = "Option::is_none")]
-    pub command_read_timeout: Option<i64>,
-    #[serde(
-        rename = "commandWriteTimeout",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub command_write_timeout: Option<i64>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<String>,
-    #[serde(rename = "functionName", skip_serializing_if = "Option::is_none")]
-    pub function_name: Option<String>,
-    #[serde(
-        rename = "maxCommandExecutionTime",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub max_command_execution_time: Option<i64>,
-    #[serde(rename = "poolSize", skip_serializing_if = "Option::is_none")]
-    pub pool_size: Option<i64>,
-    #[serde(rename = "returnName", skip_serializing_if = "Option::is_none")]
-    pub return_name: Option<String>,
-    #[serde(rename = "returnType", skip_serializing_if = "Option::is_none")]
-    pub return_type: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub runtime: Option<UdfRuntime>,
-    #[serde(rename = "sandboxType", skip_serializing_if = "Option::is_none")]
-    pub sandbox_type: Option<UdfSandboxType>,
-    #[serde(rename = "sandboxVersion", skip_serializing_if = "Option::is_none")]
-    pub sandbox_version: Option<UdfSandboxVersion>,
-    #[serde(rename = "sendChunkHeader", skip_serializing_if = "Option::is_none")]
-    pub send_chunk_header: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<UdfStatus>,
-    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<UdfType>,
-    #[serde(rename = "updatedAt", skip_serializing_if = "Option::is_none")]
-    pub updated_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub version: Option<i64>,
-}
-
-/// An argument returned in a UDF response.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfArgumentResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<String>,
-}
-
-/// Inline enum for `Udf.runtime`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfRuntime {
-    #[serde(rename = "python3.11")]
-    #[default]
-    Python3_11,
-    #[serde(rename = "native")]
-    Native,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfRuntime {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Python3_11 => write!(f, "python3.11"),
-            Self::Native => write!(f, "native"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Udf.sandboxType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfSandboxType {
-    #[serde(rename = "basic")]
-    #[default]
-    Basic,
-    #[serde(rename = "netenable")]
-    Netenable,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfSandboxType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Basic => write!(f, "basic"),
-            Self::Netenable => write!(f, "netenable"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Udf.sandboxVersion`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfSandboxVersion {
-    #[serde(rename = "v1")]
-    #[default]
-    V1,
-    #[serde(rename = "v2")]
-    V2,
-    #[serde(rename = "v3")]
-    V3,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfSandboxVersion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::V1 => write!(f, "v1"),
-            Self::V2 => write!(f, "v2"),
-            Self::V3 => write!(f, "v3"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Udf.status`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfStatus {
-    #[serde(rename = "building")]
-    #[default]
-    Building,
-    #[serde(rename = "error")]
-    Error,
-    #[serde(rename = "ready")]
-    Ready,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Building => write!(f, "building"),
-            Self::Error => write!(f, "error"),
-            Self::Ready => write!(f, "ready"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Udf.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfType {
-    #[serde(rename = "executable")]
-    #[default]
-    Executable,
-    #[serde(rename = "executable_pool")]
-    ExecutablePool,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Executable => write!(f, "executable"),
-            Self::ExecutablePool => write!(f, "executable_pool"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `UdfCreateRequest` - one of multiple variants.
-///
-/// Dispatched on the `type` field; the raw-value dispatch preserves unknown
-/// variants and prevents overlapping request shapes from being misrouted.
-#[derive(Debug, Clone, PartialEq, Serialize)]
-#[serde(untagged)]
-pub enum UdfCreateRequest {
-    UdfCreateRequestV1(UdfCreateRequestV1),
-    UdfCreateRequestV2(UdfCreateRequestV2),
-    /// Catch-all for unknown or newly-added values.
-    Unknown(serde_json::Value),
-}
-
-discriminated_union! {
-    UdfCreateRequest, "type" {
-        "executable" => UdfCreateRequestV1,
-        "executable_pool" => UdfCreateRequestV2,
-    }
-}
-
-impl std::fmt::Display for UdfCreateRequest {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::UdfCreateRequestV1(_) => write!(f, "UdfCreateRequestV1"),
-            Self::UdfCreateRequestV2(_) => write!(f, "UdfCreateRequestV2"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// The `executable` variant of [`UdfCreateRequest`].
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfCreateRequestV1 {
-    pub arguments: Vec<UdfArgument>,
-    #[serde(rename = "commandReadTimeout", skip_serializing_if = "Option::is_none")]
-    pub command_read_timeout: Option<i64>,
-    #[serde(
-        rename = "commandWriteTimeout",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub command_write_timeout: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<String>,
-    #[serde(rename = "functionName")]
-    pub function_name: String,
-    #[serde(
-        rename = "maxCommandExecutionTime",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub max_command_execution_time: Option<i64>,
-    #[serde(rename = "poolSize", skip_serializing_if = "Option::is_none")]
-    pub pool_size: Option<()>,
-    #[serde(rename = "returnName", skip_serializing_if = "Option::is_none")]
-    pub return_name: Option<String>,
-    #[serde(rename = "returnType")]
-    pub return_type: String,
-    pub runtime: UdfRuntime,
-    #[serde(rename = "sandboxType", skip_serializing_if = "Option::is_none")]
-    pub sandbox_type: Option<UdfSandboxType>,
-    #[serde(rename = "sandboxVersion", skip_serializing_if = "Option::is_none")]
-    pub sandbox_version: Option<UdfSandboxVersion>,
-    #[serde(rename = "sendChunkHeader", skip_serializing_if = "Option::is_none")]
-    pub send_chunk_header: Option<bool>,
-    #[serde(rename = "type")]
-    pub r#type: UdfCreateRequestV1Type,
-    #[serde(rename = "uploadId")]
-    pub upload_id: String,
-}
-
-/// Inline enum for `UdfCreateRequestV1.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfCreateRequestV1Type {
-    #[serde(rename = "executable")]
-    #[default]
-    Executable,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfCreateRequestV1Type {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Executable => write!(f, "executable"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// The `executable_pool` variant of [`UdfCreateRequest`].
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfCreateRequestV2 {
-    pub arguments: Vec<UdfArgument>,
-    #[serde(rename = "commandReadTimeout", skip_serializing_if = "Option::is_none")]
-    pub command_read_timeout: Option<i64>,
-    #[serde(
-        rename = "commandWriteTimeout",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub command_write_timeout: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<String>,
-    #[serde(rename = "functionName")]
-    pub function_name: String,
-    #[serde(
-        rename = "maxCommandExecutionTime",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub max_command_execution_time: Option<i64>,
-    #[serde(rename = "poolSize", skip_serializing_if = "Option::is_none")]
-    pub pool_size: Option<i64>,
-    #[serde(rename = "returnName", skip_serializing_if = "Option::is_none")]
-    pub return_name: Option<String>,
-    #[serde(rename = "returnType")]
-    pub return_type: String,
-    pub runtime: UdfRuntime,
-    #[serde(rename = "sandboxType", skip_serializing_if = "Option::is_none")]
-    pub sandbox_type: Option<UdfSandboxType>,
-    #[serde(rename = "sandboxVersion", skip_serializing_if = "Option::is_none")]
-    pub sandbox_version: Option<UdfSandboxVersion>,
-    #[serde(rename = "sendChunkHeader", skip_serializing_if = "Option::is_none")]
-    pub send_chunk_header: Option<bool>,
-    #[serde(rename = "type")]
-    pub r#type: UdfCreateRequestV2Type,
-    #[serde(rename = "uploadId")]
-    pub upload_id: String,
-}
-
-/// Inline enum for `UdfCreateRequestV2.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfCreateRequestV2Type {
-    #[serde(rename = "executable_pool")]
-    #[default]
-    ExecutablePool,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfCreateRequestV2Type {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ExecutablePool => write!(f, "executable_pool"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `UdfListResponse` from the ClickHouse Cloud API.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfListResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub items: Option<Vec<Udf>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pagination: Option<Pagination>,
-}
-
-/// `UdfUploadSession` from the ClickHouse Cloud API.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfUploadSession {
-    #[serde(rename = "expiresAt", skip_serializing_if = "Option::is_none")]
-    pub expires_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(rename = "uploadId", skip_serializing_if = "Option::is_none")]
-    pub upload_id: Option<String>,
-    #[serde(rename = "uploadUrl", skip_serializing_if = "Option::is_none")]
-    pub upload_url: Option<String>,
-}
-
-/// `UdfVersionCreateRequest` - one of multiple variants.
-///
-/// Dispatched on the `type` field; the raw-value dispatch preserves unknown
-/// variants and prevents overlapping request shapes from being misrouted.
-#[derive(Debug, Clone, PartialEq, Serialize)]
-#[serde(untagged)]
-pub enum UdfVersionCreateRequest {
-    UdfVersionCreateRequestV1(UdfVersionCreateRequestV1),
-    UdfVersionCreateRequestV2(UdfVersionCreateRequestV2),
-    /// Catch-all for unknown or newly-added values.
-    Unknown(serde_json::Value),
-}
-
-discriminated_union! {
-    UdfVersionCreateRequest, "type" {
-        "executable" => UdfVersionCreateRequestV1,
-        "executable_pool" => UdfVersionCreateRequestV2,
-    }
-}
-
-impl std::fmt::Display for UdfVersionCreateRequest {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::UdfVersionCreateRequestV1(_) => write!(f, "UdfVersionCreateRequestV1"),
-            Self::UdfVersionCreateRequestV2(_) => write!(f, "UdfVersionCreateRequestV2"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// The `executable` variant of [`UdfVersionCreateRequest`].
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfVersionCreateRequestV1 {
-    pub arguments: Vec<UdfArgument>,
-    #[serde(rename = "commandReadTimeout", skip_serializing_if = "Option::is_none")]
-    pub command_read_timeout: Option<i64>,
-    #[serde(
-        rename = "commandWriteTimeout",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub command_write_timeout: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<String>,
-    #[serde(
-        rename = "maxCommandExecutionTime",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub max_command_execution_time: Option<i64>,
-    #[serde(rename = "poolSize", skip_serializing_if = "Option::is_none")]
-    pub pool_size: Option<()>,
-    #[serde(rename = "returnName", skip_serializing_if = "Option::is_none")]
-    pub return_name: Option<String>,
-    #[serde(rename = "returnType")]
-    pub return_type: String,
-    pub runtime: UdfRuntime,
-    #[serde(rename = "sandboxType", skip_serializing_if = "Option::is_none")]
-    pub sandbox_type: Option<UdfSandboxType>,
-    #[serde(rename = "sandboxVersion", skip_serializing_if = "Option::is_none")]
-    pub sandbox_version: Option<UdfSandboxVersion>,
-    #[serde(rename = "sendChunkHeader", skip_serializing_if = "Option::is_none")]
-    pub send_chunk_header: Option<bool>,
-    #[serde(rename = "type")]
-    pub r#type: UdfVersionCreateRequestV1Type,
-    #[serde(rename = "uploadId")]
-    pub upload_id: String,
-}
-
-/// Inline enum for `UdfVersionCreateRequestV1.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfVersionCreateRequestV1Type {
-    #[serde(rename = "executable")]
-    #[default]
-    Executable,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfVersionCreateRequestV1Type {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Executable => write!(f, "executable"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// The `executable_pool` variant of [`UdfVersionCreateRequest`].
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfVersionCreateRequestV2 {
-    pub arguments: Vec<UdfArgument>,
-    #[serde(rename = "commandReadTimeout", skip_serializing_if = "Option::is_none")]
-    pub command_read_timeout: Option<i64>,
-    #[serde(
-        rename = "commandWriteTimeout",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub command_write_timeout: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<String>,
-    #[serde(
-        rename = "maxCommandExecutionTime",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub max_command_execution_time: Option<i64>,
-    #[serde(rename = "poolSize", skip_serializing_if = "Option::is_none")]
-    pub pool_size: Option<i64>,
-    #[serde(rename = "returnName", skip_serializing_if = "Option::is_none")]
-    pub return_name: Option<String>,
-    #[serde(rename = "returnType")]
-    pub return_type: String,
-    pub runtime: UdfRuntime,
-    #[serde(rename = "sandboxType", skip_serializing_if = "Option::is_none")]
-    pub sandbox_type: Option<UdfSandboxType>,
-    #[serde(rename = "sandboxVersion", skip_serializing_if = "Option::is_none")]
-    pub sandbox_version: Option<UdfSandboxVersion>,
-    #[serde(rename = "sendChunkHeader", skip_serializing_if = "Option::is_none")]
-    pub send_chunk_header: Option<bool>,
-    #[serde(rename = "type")]
-    pub r#type: UdfVersionCreateRequestV2Type,
-    #[serde(rename = "uploadId")]
-    pub upload_id: String,
-}
-
-/// Inline enum for `UdfVersionCreateRequestV2.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UdfVersionCreateRequestV2Type {
-    #[serde(rename = "executable_pool")]
-    #[default]
-    ExecutablePool,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UdfVersionCreateRequestV2Type {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ExecutablePool => write!(f, "executable_pool"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `UdfVersionListResponse` from the ClickHouse Cloud API.
-///
-/// Used in response position only: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UdfVersionListResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub items: Option<Vec<Udf>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pagination: Option<Pagination>,
-}
-
-/// Standard API response wrapper.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-#[serde(from = "crate::serde_helpers::ApiResponseWire<T>")]
-pub struct ApiResponse<T> {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none", rename = "requestId")]
-    pub request_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub result: Option<T>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
 }
 
 impl Default for BackupBucket {
